+++ src/dataframe/dataframe.jl
@@ -966,8 +966,14 @@ function Base.delete!(df::DataFrame, inds)
         throw(BoundsError(df, (inds, :)))
     end
 
+    # workaround https://github.com/JuliaLang/julia/pull/41646
+    if VERSION <= v"1.6.2" && inds isa UnitRange{<:Integer}
+        inds = collect(inds)
+    end
+
     # we require ind to be stored and unique like in Base
     # otherwise an error will be thrown and the data frame will get corrupted
+
     return _delete!_helper(df, inds)
 end
 
@@ -976,6 +982,10 @@ function Base.delete!(df::DataFrame, inds::AbstractVector{Bool})
         throw(BoundsError(df, (inds, :)))
     end
     drop = _findall(inds)
+    # workaround https://github.com/JuliaLang/julia/pull/41646
+    if VERSION <= v"1.6.2" && drop isa UnitRange{<:Integer}
+        drop = collect(drop)
+    end
     return _delete!_helper(df, drop)
 end